Split out aggregation from its executable
